set up authenticate

+++ user/internal/application/common/command/refresh_result.go
@@ -3,6 +3,6 @@ package common
 type RefreshTokenCommandResult struct {
 	AccessToken           string `json:"access_token"`
 	RefreshToken          string `json:"refresh_token"`
-	AccessTokenExpiresAt  int    `json:"access_token_expires_at"`
-	RefreshTokenExpiresAt int    `json:"refresh_token_expires_at"`
+	AccessTokenExpiresAt  int64  `json:"access_token_expires_at"`
+	RefreshTokenExpiresAt int64  `json:"refresh_token_expires_at"`
 }
